Moved indicator into browser utils
