+++ apps/browser/src/util/browser.ts
@@ -1,4 +1,5 @@
 import App from "@browser/util/app";
+import Registry from "@browser/state/registry";
 
 type Data = {
   [key: string]: any;
@@ -128,5 +129,37 @@ export class Storage {
         res(data[key]);
       });
     });
+  }
+}
+
+export class Indicator {
+  static hide() {
+    if (!chrome.browserAction) return { success: false };
+    chrome.browserAction.setIcon({
+      path: {
+        "64": chrome.runtime.getURL("assets/icon/64.png"),
+        "128": chrome.runtime.getURL("assets/icon/128.png"),
+        "256": chrome.runtime.getURL("assets/icon/256.png"),
+      },
+    });
+    return { success: true };
+  }
+
+  static show() {
+    if (!chrome.browserAction) return { success: false };
+    chrome.browserAction.setIcon({
+      path: {
+        "64": chrome.runtime.getURL("assets/indicator/64.png"),
+        "128": chrome.runtime.getURL("assets/indicator/128.png"),
+        "256": chrome.runtime.getURL("assets/indicator/256.png"),
+      },
+    });
+    return { success: true };
+  }
+
+  static check() {
+    const state = Registry.store.getState();
+    if (state.comment.total > 0) chrome.runtime.sendMessage("INDICATOR:SHOW");
+    else chrome.runtime.sendMessage("INDICATOR:HIDE");
   }
 }
